backend/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy

youtube-transcript-api

groq

python-dotenv
pydantic
huggingface-hub
